+++ test/str_test.py
@@ -116,8 +116,9 @@ class TestStringRepresentations(unittest.TestCase):
         cover = Cover(
             xknx,
             name='Rolladen',
-            group_address_long='1/2/3',
-            group_address_short='1/2/4',
+            group_address_long='1/2/2',
+            group_address_short='1/2/3',
+            group_address_stop='1/2/4',
             group_address_position='1/2/5',
             group_address_position_state='1/2/6',
             group_address_angle='1/2/7',
@@ -126,9 +127,10 @@ class TestStringRepresentations(unittest.TestCase):
             travel_time_up=10)
         self.assertEqual(
             str(cover),
-            '<Cover name="Rolladen" updown="GroupAddress("1/2/3")/None/None/None" step="GroupAddress("1/2/4")/None/None/None" position="Group'
-            'Address("1/2/5")/GroupAddress("1/2/6")/None/None" angle="GroupAddress("1/2/7")/GroupAddress("1/2/8")/None/None" travel_time_down="8'
-            '" travel_time_up="10" />')
+            '<Cover name="Rolladen" updown="GroupAddress("1/2/2")/None/None/None" step="GroupAddress("1/2/3")/None/None/None" '
+            'stop="GroupAddress("1/2/4")/None/None/None" position="GroupAddress("1/2/5")/GroupAddress("1/2/6")/None/None" '
+            'angle="GroupAddress("1/2/7")/GroupAddress("1/2/8")/None/None" '
+            'travel_time_down="8" travel_time_up="10" />')
 
     def test_fan(self):
         """Test string representation of fan object."""